+++ demos/web-editor/web/src/worker.ts
@@ -11,7 +11,7 @@ import {
 
 import { RENDER_SIZE } from "./constants";
 
-var fidget: any = null;
+import * as fidget from "../../crate/pkg/fidget_wasm_demo";
 
 class Worker {
   render(s: ShapeRequest) {
@@ -19,15 +19,18 @@ class Worker {
     let out: Uint8Array;
     switch (s.mode) {
       case RenderMode.Bitmap: {
-        out = fidget.render_region_2d(shape, RENDER_SIZE);
+        const camera = fidget.JsCamera2.deserialize(s.camera);
+        out = fidget.render_region_2d(shape, RENDER_SIZE, camera);
         break;
       }
       case RenderMode.Heightmap: {
-        out = fidget.render_region_heightmap(shape, RENDER_SIZE);
+        const camera = fidget.JsCamera3.deserialize(s.camera);
+        out = fidget.render_region_heightmap(shape, RENDER_SIZE, camera);
         break;
       }
       case RenderMode.Normals: {
-        out = fidget.render_region_normals(shape, RENDER_SIZE);
+        const camera = fidget.JsCamera3.deserialize(s.camera);
+        out = fidget.render_region_normals(shape, RENDER_SIZE, camera);
         break;
       }
     }
@@ -61,7 +64,6 @@ class Worker {
 }
 
 async function run() {
-  fidget = await import("../../crate/pkg/fidget_wasm_demo.js")!;
   await fidget.default();
   await fidget.initThreadPool(navigator.hardwareConcurrency);